% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_empatica__main.R
\name{read_empatica_avro}
\alias{read_empatica_avro}
\title{Read an Empatica raw data file}
\usage{
read_empatica_avro(f, verbose = FALSE)
}
\arguments{
\item{f}{character. Path to the file}

\item{verbose}{logical. Print updates to console?}
}
\value{
a list of formatted data elements
}
\description{
Read an Empatica raw data file
}
\examples{

f <- system.file("extdata/test_file.avro", package = "empatica")
head( read_empatica_avro(f) )

}
